feat(services): payment services both now use Stripe. Logic re-done

+++ src/main/java/org/example/gym_shop_2026/services/paymentMethodService.java
@@ -1,5 +1,11 @@
 package org.example.gym_shop_2026.services;
 
+import com.stripe.exception.StripeException;
+import com.stripe.model.Customer;
+import com.stripe.model.PaymentMethod;
+import com.stripe.param.CustomerCreateParams;
+import com.stripe.param.PaymentMethodAttachParams;
+import org.example.gym_shop_2026.entities.User;
 import org.example.gym_shop_2026.entities.paymentMethod;
 import org.example.gym_shop_2026.persistence.paymentMethodDAO;
 import org.springframework.stereotype.Service;
@@ -10,30 +16,48 @@ import java.util.List;
 @Service
 public class paymentMethodService {
     private final paymentMethodDAO paymentMethodDao;
+    private final UserService userService;
 
-    public paymentMethodService(paymentMethodDAO paymentMethodDao) {
+    public paymentMethodService(paymentMethodDAO paymentMethodDao, UserService userService) {
         this.paymentMethodDao = paymentMethodDao;
+        this.userService = userService;
     }
 
-    public boolean addPaymentMethod(paymentMethod method) throws SQLException {
-        if (paymentMethodDao.isPaymentTokenUnique(method.getProcessorToken())) {
-            return paymentMethodDao.insertPaymentMethod(method);
+    public boolean addPaymentMethod(paymentMethod method) throws SQLException, StripeException {
+        User user = userService.getUserById(method.getUserId());
+        String stripeCustomerId = user.getStripeCustomerId();
+
+        if (stripeCustomerId == null || stripeCustomerId.isEmpty()) {
+            CustomerCreateParams customerParams = CustomerCreateParams.builder()
+                    .setEmail(user.getEmail())
+                    .setName(user.getFullName())
+                    .build();
+            Customer customer = Customer.create(customerParams);
+            stripeCustomerId = customer.getId();
+
+            userService.updateStripeCustomerId(user.getUser_id(), stripeCustomerId);
         }
-        return false;
+
+        PaymentMethod pm = PaymentMethod.retrieve(method.getStripePaymentMethodId());
+        pm.attach(PaymentMethodAttachParams.builder()
+                .setCustomer(stripeCustomerId)
+                .build());
+
+        return paymentMethodDao.insertPaymentMethod(method);
     }
 
-    /**
-     * Calls insertPaymentMethodTesting() method from {@link paymentMethodDAO}
-     * @param method {@link paymentMethod} to insert
-     * @return True if dao method successful
-     * @throws SQLException If dao method fails
-     */
-    public boolean addPaymentMethodTest(paymentMethod method) throws SQLException {
-        if (paymentMethodDao.isPaymentTokenUnique(method.getProcessorToken())) {
-            return paymentMethodDao.insertPaymentMethodTesting(method);
-        }
-        return false;
-    }
+//    /**
+//     * Calls insertPaymentMethodTesting() method from {@link paymentMethodDAO}
+//     * @param method {@link paymentMethod} to insert
+//     * @return True if dao method successful
+//     * @throws SQLException If dao method fails
+//     */
+//    public boolean addPaymentMethodTest(paymentMethod method) throws SQLException {
+//        if (paymentMethodDao.isPaymentTokenUnique(method.getProcessorToken())) {
+//            return paymentMethodDao.insertPaymentMethodTesting(method);
+//        }
+//        return false;
+//    }
 
     public List<paymentMethod> getAllUserMethods(int userId) throws SQLException {
         return paymentMethodDao.findAllMethodsByUserId(userId);
@@ -59,8 +83,15 @@ public class paymentMethodService {
         return paymentMethodDao.updateMethodValidity(methodId, isValid);
     }
 
-    public paymentMethod deletePaymentMethod(int id) throws SQLException {
-        return paymentMethodDao.deleteMethodById(id);
+    public paymentMethod deletePaymentMethod(int id) throws SQLException, StripeException {
+        paymentMethod method = paymentMethodDao.findMethodById(id);
+        if (method != null) {
+            PaymentMethod pm = PaymentMethod.retrieve(method.getStripePaymentMethodId());
+            pm.detach();
+
+            return paymentMethodDao.deleteMethodById(id);
+        }
+        return null;
     }
 
     /**
